# chore: status command performance improvements, remove unused api commands