Reference: https://linear.app/speakeasy/issue/SPE-4008/feature-search-workspace-event-api-success-and-result-limit-filtering
Reference: https://github.com/speakeasy-api/speakeasy-registry/pull/3240

This upgrades `github.com/speakeasy-api/speakeasy-client-sdk-go/v3` to latest, which contains unused operation cleanup in the Speakeasy API which was propagated to the Go SDK. Removes associated code.

This change also enhances the `status` command to use new search workspace event limit and success filtering to increase the command's performance when searching target generate/success events. It will first fetch the last event, and if that event was successful, will also use that for the last successful event rather than making an extra API call. Each target's data is also fetched concurrently.

Previously:

```
# bflad-testing (3 targets)
speakeasy status  1.91s user 0.08s system 47% cpu 4.223 total

# dub (5 targets)
speakeasy status  2.26s user 0.10s system 31% cpu 7.604 total
```

Now:

```
# bflad-testing (3 targets)
~/go/bin/speakeasy status  1.79s user 0.08s system 89% cpu 2.080 total

# dub (5 targets)
~/go/bin/speakeasy status  1.74s user 0.07s system 76% cpu 2.369 total
```